+++ packages/styles/src/packages/label/index.ts
@@ -1,0 +1,5 @@
+import { style } from "./label.style.css";
+
+export const label = {
+  style,
+};